requirements.txt
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cryptography==43.0.3
cssbeautifier==1.15.1
Django==5.1.2
django-allauth==65.1.0
django-browser-reload==1.17.0
django-extensions==3.2.3
django-tailwind==3.8.0
djlint==1.35.2
EditorConfig==0.12.4
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.9.25
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
pathspec==0.12.1
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.3
rsa==4.9
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
text-unidecode==1.3
tqdm==4.66.6
types-python-dateutil==2.9.0.20241003
uritemplate==4.1.1
urllib3==2.2.3
